Countdown timer plus other features

+++ src/global.hpp
@@ -1,7 +1,7 @@
 #ifndef _GLOBAL_HEADER_STOCKTRACKER_
 #define _GLOBAL_HEADER_STOCKTRACKER_
 
-using time_record_t = long long;
+using time_record_t = qint64;
 using price_t       = double;
 using percentage_t  = double;
 using volume_t      = long long;
